added interval and target options
if the target option is used, the application will exit if/when the idle
time reaches the supplied target value, in milliseconds.

target mode is silent by default, toggable with the -v switch.

to change how often the idle value is checked, use the -i switch. it
accepts a parameter in milliseconds.